voxel

+++ src/moai-sim/MOAIIndexBuffer.cpp
@@ -101,6 +101,7 @@ MOAIIndexBuffer::~MOAIIndexBuffer () {
 void MOAIIndexBuffer::OnBind () {
 
 	if ( this->mGLBufferID ) {
+		//printf("OnBind %d\n",this->mGLBufferID);
 		zglBindBuffer ( ZGL_BUFFER_TARGET_ELEMENT_ARRAY, this->mGLBufferID );
 	}
 }
@@ -123,7 +124,7 @@ void MOAIIndexBuffer::OnCreate () {
 		if ( this->mGLBufferID ) {
 		
 			zglBindBuffer ( ZGL_BUFFER_TARGET_ELEMENT_ARRAY, this->mGLBufferID );
-			zglBufferData ( ZGL_BUFFER_TARGET_ELEMENT_ARRAY, this->mIndexCount * sizeof ( u16 ), this->mBuffer, this->mHint );
+			zglBufferData ( ZGL_BUFFER_TARGET_ELEMENT_ARRAY, this->mIndexCount * sizeof ( u32 ), this->mBuffer, this->mHint );
 		}
 	}
 }
@@ -171,13 +172,13 @@ void MOAIIndexBuffer::ReserveIndices ( u32 indexCount ) {
 	this->Clear ();
 	
 	this->mIndexCount = indexCount;
-	this->mBuffer = ( u16* )malloc ( indexCount * sizeof ( u16 ));
+	this->mBuffer = ( u32* )malloc ( indexCount * sizeof ( u32 ));
 	
 	this->Load ();
 }
 
 //----------------------------------------------------------------//
-void MOAIIndexBuffer::SetIndex ( u32 idx, u16 value ) {
+void MOAIIndexBuffer::SetIndex ( u32 idx, u32 value ) {
 
 	if ( idx < this->mIndexCount ) {
 		this->mBuffer [ idx ] = value;
