+++ src/moai-sim/MOAIIndexBuffer.h
@@ -17,7 +17,7 @@ class MOAIIndexBuffer :
 	public MOAIGfxResource {
 private:
 
-	u16*	mBuffer;
+	u32*	mBuffer;
 	u32		mIndexCount;
 	
 	u32		mGLBufferID;
@@ -51,7 +51,7 @@ public:
 	void		RegisterLuaClass		( MOAILuaState& state );
 	void		RegisterLuaFuncs		( MOAILuaState& state );
 	void		ReserveIndices			( u32 indexCount );
-	void		SetIndex				( u32 idx, u16 value );
+	void		SetIndex				( u32 idx, u32 value );
 };
 
 #endif
